[Analysis API] Fix overridden declaration computation for fake overrides
According to the commit history, the previous behavior took place
likely because of issues with implementing the override/implement
dialog. However, it broke FIR implementations of
'getAllOverriddenSymbols()' and 'getDirectlyOverriddenSymbols()' in case
if a passed callable is a declaration-side substitution override.

According to my tests, the mentioned functionality works as before
after the change. In addition, there are several related tests in
the IntelliJ Kotlin plugin that still pass:

- FirOverrideImplementTest.testComplexMultiOverride()
- FirOverrideImplementTest.testGenerateMulti()
- FirOverrideImplementTest.testGenericClass()

Note that in the current implementation, fake overrides are always
unwrapped according to the KDoc of API methods.

^KT-62405 Fixed