(TODO) IR: simplify fake override modality computation algorithm
Do not collect all overridden declarations because they are not needed.
To compute the modality of a fake override, we can only consider
modalities of immediate overridden declarations. However we must remove
those immediate overridden declarations which are overridden by some
other immediate overridden declaration. This is needed because an
abstract declaration might override an open declaration, in which case
the resulting modality should be abstract.

This change incidentally fixes KT-65025, however the root problem of
FakeOverrideGenerator not supporting RenamedForOverride is likely still
relevant.

 #KT-65025 Fixed