[Wasm] Use general KClass logic for primitive types
K/Wasm was reusing K/JS logic for getting KClass of primitive types.
This is not necessary in K/Wasm as all primitive types
have general runtime type information.

Remove special KClasses for Nothing and Error types.
For this Nothing type now has to be a real class,
without @ExcludeFromCodegen. DCE already eliminates this class when
 used as `Nothing`. A case for `Nothing?` was added as well.

This commit fixes the issue of a missing Short
case in getKClassFromExpression

^KT-65210: Fixed