[Analysis API] Support type parameter types in type code fragments
The initial implementation wrapped a type reference to the
'null is <type>' expression. It worked for class and interface types,
however for non-reified type parameters there was an error
'FirErrors.CANNOT_CHECK_FOR_ERASED'.

The new implementation replaces the instance check with the function
literal 'fun(value: <type>) {}', which seems to have no problems with
generic types.

^KT-65552 Fixed