[IR] Interpret JS constants the same way as K1 does
Earlier, it was decided that IR interpreter will be the main source
of constant evaluation in K2. With that, there were implemented a
couple of fixes in evaluation of JS constants. For example, in K1
expression like `1.0.toString()` will be evaluated into `1.0` although,
by JS rules, it must be evaluated into `1`

But for many reasons, we actually need to be able to evaluate
constants on FIR tree. With that decision, we need to match IR and FIR
evaluation results. So, for now, IR interpreter will evaluate constants
on JS platform using K1 rules.