Rename all visitExtensions methods on Km nodes to getExtension
and document @InternalExtensionsApi properly.

These functions were previously a part of the deprecated and removed Visitors API.
However, it is necessary for the mechanism of platform extensions, and
it is marked now as @InternalExtensionsApi.
It is logical to rename it to get rid of 'visit' part and to explain
that it returns only one extension.

See also: KT-63156