K2: ignore candidate receiver with error in CheckExtensionReceiver
If a candidate's receiver is resolved to a error type, we consider
any other type as a subtype of it and therefore may select
any candidate that we happen to find in a scope.
In particular, in the case of scripts, the receiver candidate
resolved to a cycle was accepted as a receiver to an invoke on a
random class from stdlib.
The fix rejects the candidate in this case, allowing the tower to find
the correct one.

#KT-64241 fixed