[FIR] Fix giant recursive type produced by CST
The symptom of the problem, compiler/fir/analysis-tests/testData/resolve
/inference/onlyInputTypesCapturedTypeWithRecursiveBounds.fir.txt was
already accidentally fixed by 81414d75 (KT-65596).
However, we want to return to the new behavior (KT-59138) in the future,
therefore, we also want the cause of the problem fixed.

The fix was manually verified by cherry-picking it on top of the last
commit before 81414d75.

#KT-65704 Fixed